Add Full Functionality For Tests Management

+++ DVLDBusinessLayer/LicenseClasses.cs
@@ -6,15 +6,14 @@ namespace DVLDBusinessLayer
 {
     public class clsLicenseClasses
     {
-        public int ID { get; set; }
+        public byte ID { get; set; }
         public string ClassName { get; set; }
-
         public static DataTable GetLicenseClassesNames()
         {
             return clsLicenseClassesData.GetLicenseClassesNames();
         }
 
-        public static int GetLicenseClassID(string LicenseClassName)
+        public static byte GetLicenseClassID(string LicenseClassName)
         {
             return clsLicenseClassesData.GetLicenseClassID(LicenseClassName);
         }
